% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree-methods.R
\name{ancestorDistance}
\alias{ancestorDistance}
\title{distance between a descendant and an ancestor}
\usage{
ancestorDistance(procPed, a, d)
}
\arguments{
\item{procPed}{pedigree that has been through \code{processPedigree}}

\item{a}{ancestor subject}

\item{d}{descendant subject}
}
\value{
minimum distance (number generations) between a and d
}
\description{
distance between a descendant and an ancestor
}
\keyword{internal}

