% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree-methods.R
\name{areMating}
\alias{areMating}
\title{determine if two subjects have a child together}
\usage{
areMating(procPed, f1, f2)
}
\arguments{
\item{procPed}{pedigree that has been through \code{processPedigree}}

\item{f1}{subject 1}

\item{f2}{subject 2}
}
\value{
true if both subjects share a child
}
\description{
determine if two subjects have a child together
}
\keyword{internal}

