% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing.R
\name{checkArgs}
\alias{checkArgs}
\title{check arguments provided to RVsharing for validty}
\usage{
checkArgs(alleleFreq, kinshipCoeff, nSim, founderDist)
}
\arguments{
\item{alleleFreq}{allele frequency among the founders}

\item{kinshipCoeff}{mean kinship coefficient among the founders}

\item{nSim}{number of simulations used in monte carlo calculation}

\item{founderDist}{custom distribution among founders - only used
when simulating probability with nSim}
}
\value{
throws error if arguments invalid
}
\description{
verifies that arguments are valid, throws an error
if they are not
}
\keyword{internal}
