% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.esllasso.R
\name{coef.esllasso}
\alias{coef.esllasso}
\title{Provides estimated coefficients from a fitted "esllasso" object.}
\usage{
\method{coef}{esllasso}(object, ...)
}
\arguments{
\item{object}{An "\code{esllasso}" project.}

\item{...}{Other arguments.}
}
\value{
A list consisting of the intercept and regression coefficients of the fitted model.
}
\description{
This function provides estimated
coefficients from a fitted "\code{esllasso}" object.
}
