% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.robustlm.R
\name{coef.robustlm}
\alias{coef.robustlm}
\title{Provides estimated coefficients from a fitted "robustlm" object.}
\usage{
\method{coef}{robustlm}(object, ...)
}
\arguments{
\item{object}{An "\code{robustlm}" project.}

\item{...}{Other arguments.}
}
\value{
A list consisting of the intercept and regression coefficients of the fitted model.
}
\description{
This function provides estimated
coefficients from a fitted "\code{robustlm}" object.
}
