% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.rvs.R
\name{coef.rvs}
\alias{coef.rvs}
\title{Provides estimated coefficients from a fitted "rvs" object.}
\usage{
\method{coef}{rvs}(object, ...)
}
\arguments{
\item{object}{An "\code{rvs}" project.}

\item{...}{Other arguments.}
}
\value{
A list consisting of the intercept and regression coefficients of the fitted model.
}
\description{
This function provides estimated
coefficients from a fitted "\code{rvs}" object.
}
