% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedFoundersCorrection.R
\name{computePFU}
\alias{computePFU}
\title{computation of P[FjU] using equation 21 of Bureau et al.}
\usage{
computePFU(nf, theta, ord = 2)
}
\arguments{
\item{nf}{number of founders of the pedigree}

\item{theta}{value of the parameter of the polynomial distribution}

\item{ord}{order of the polynomial approximation to the distribtion of the
number of distinct alleles in the founders (noted d in Bureay et al.).
Must be <= 5}
}
\value{
P[FjU] (scalar)
}
\description{
computation of P[FjU] using equation 21 of Bureau et al.
}
\keyword{internal}

