% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedFoundersCorrection.R
\name{computePhiVec}
\alias{computePhiVec}
\title{\code{computePhiVec} the vector of expected phi_a for nf founders for
 numbers of distinct alleles a from amin to 2*nf-1}
\usage{
computePhiVec(nf, amin = 2 * nf - 2)
}
\description{
\code{computePhiVec} the vector of expected phi_a for nf founders for
 numbers of distinct alleles a from amin to 2*nf-1
}
\keyword{internal}
