% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculationsBackend.R
\name{convertMatrix}
\alias{convertMatrix}
\title{convert snpMatrix to a list of vectors of sharing}
\usage{
convertMatrix(snpMat, famIds, minorAllele)
}
\arguments{
\item{snpMat}{SnpMatrix}

\item{famIds}{family ids corresponding to rows of the snpMap}

\item{minorAllele}{vector specifying the minor allele of each variant}
}
\value{
list of boolean vectors indicating sharing pattern for each variant
}
\description{
convert snpMatrix to a list of vectors of sharing
}
