% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing2.R
\name{createNetwork}
\alias{createNetwork}
\title{\code{createNetwork} create bayesian network from pedigree}
\usage{
createNetwork(ped, prior = c(1, 1, 1))
}
\arguments{
\item{ped}{S3 Pedigree object}

\item{prior}{prior on number of alleles for founders}
}
\value{
bayesian network friom gRain package
}
\description{
\code{createNetwork} create bayesian network from pedigree
}
\keyword{internal}

