% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculationsBackend.R
\name{enrichmentPValue_R}
\alias{enrichmentPValue_R}
\title{R backend for enrichmentPValue calculation}
\usage{
enrichmentPValue_R(snpMat, famIds, sharingProbs, minorAllele,
  threshold = 0)
}
\arguments{
\item{snpMat}{SnpMatrix}

\item{sharingProbs}{vector of sharing probabilites, must be a named vector
with famid's for each probability}

\item{threshold}{minimum p-value threshold passed to multipleFamilyPValue}
}
\value{
p-value
}
\description{
R backend for enrichmentPValue calculation
}
