% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculationsBackend.R
\name{enrichmentPValue_R_Backend}
\alias{enrichmentPValue_R_Backend}
\title{R backend for enrichmentPValue calculation}
\usage{
enrichmentPValue_R_Backend(
  snpMat,
  famIds,
  sharingProbs,
  minorAllele,
  threshold = 0
)
}
\arguments{
\item{snpMat}{SnpMatrix}

\item{famIds}{family ids corresponding to rows of the snpMap}

\item{sharingProbs}{vector of sharing probabilites, must be a named vector
with famid's for each probability}

\item{minorAllele}{which variant value to count as the minor allele}

\item{threshold}{minimum p-value threshold passed to multipleFamilyPValue}
}
\value{
p-value
}
\description{
R backend for enrichmentPValue calculation
}
