% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingProbabilityCalculations.R
\name{exactSharingProb}
\alias{exactSharingProb}
\title{exact sharing probability calculation}
\usage{
exactSharingProb(procPed, alleleFreq)
}
\arguments{
\item{procPed}{pedigree that has been through processPedigree()}

\item{alleleFreq}{allele frequency among the founders}
}
\value{
sharing probability
}
\description{
Calculate the exact sharing probability given the minor allele
frequency among the founders (population).
}
\keyword{internal}
