% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVgene.R
\name{extract_carriers}
\alias{extract_carriers}
\title{extract carriers of minor allele}
\usage{
extract_carriers(ped, site, fam, type = "alleles", minor.allele = 2)
}
\arguments{
\item{ped}{pedigree coded in a ped file with either two alleles per
variant ("alleles"), or a count of one allele ("count")}

\item{site}{site where to record carriers}

\item{fam}{ID of the family for which to extract carriers}

\item{type}{representation of allele count}

\item{minor.allele}{id of minor allele}
}
\value{
carriers in ped
}
\description{
finds the carriers of the minor allele at a specified site
}
\keyword{internal}

