% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree-methods.R
\name{founderOccurence}
\alias{founderOccurence}
\title{determine if subjects are descended from founders}
\usage{
founderOccurence(procPed, subjects, founders)
}
\arguments{
\item{procPed}{pedigree that has been through \code{processPedigree}}

\item{subjects}{vector of subject ids}

\item{founders}{vector of founder ids}
}
\value{
data frame with 0/1 for if a subject if descended from founder
}
\description{
determine if subjects are descended from founders
}
\keyword{internal}
