% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloSharingProb.R
\name{GeneDrop}
\alias{GeneDrop}
\alias{GeneDropSim.allsubsets.fn}
\alias{GeneDropSim.fn}
\alias{GeneDropSimExcessSharing.fn}
\title{depreciated function}
\usage{
GeneDrop(...)

GeneDropSim.allsubsets.fn(...)

GeneDropSim.fn(...)

GeneDropSimExcessSharing.fn(...)
}
\arguments{
\item{...}{arguments to the old function}
}
\description{
This function is depreciated with version >= 2.0
 and should not be used, instead use RVsharing with nSim option
}

