% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculations.R
\name{get.psubset}
\alias{get.psubset}
\title{depreciated function}
\usage{
get.psubset(vec, not, pshare.data)
}
\arguments{
\item{vec}{a vector of names of all families where a variant is seen}

\item{not}{a vector of names of families where not all affected subjects
share the rare variant}

\item{pshare.data}{a data frame with at least two of the following columns:
pshare: vector of RV sharing probabilities
ped.tocompute.vec: vector of names of the families whose sharing 
    probability is contained in pshare. The names in the arguments
    vec and not must be found in ped.tocompute.vec}
}
\description{
This function is depreciated with version >= 2.0
 and should not be used, instead use multipleFamilyPValue
}

