% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedFoundersCorrection.R
\name{inferNumAlleles}
\alias{inferNumAlleles}
\title{\code{inferNumAlleles} Returns the most likely number of distinct alleles
 among nf founders based on mean estimated kinship phi}
\usage{
inferNumAlleles(phi, nf)
}
\description{
\code{inferNumAlleles} Returns the most likely number of distinct alleles
 among nf founders based on mean estimated kinship phi
}
\keyword{internal}

