% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedFoundersCorrection.R
\name{inferTheta}
\alias{inferTheta}
\title{\code{inferTheta} solve the parameter theta for polynomial approximation
 of the distribution of the number of distinct alleles. This is a general
function for polynomials of order 2 to 5.}
\usage{
inferTheta(phi, phiVec)
}
\arguments{
\item{phi}{the mean estimated kinship between founders}

\item{phiVec}{contains phi_a for a = 2*nf-ord to 2*nf-1, where
ord must be between 2 and 5}
}
\value{
real roots of the polynomial approximation
}
\description{
\code{inferTheta} solve the parameter theta for polynomial approximation
 of the distribution of the number of distinct alleles. This is a general
function for polynomials of order 2 to 5.
}
\keyword{internal}

