% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree-methods.R
\name{isDescendant}
\alias{isDescendant}
\title{determine if one subject is a descendant of another}
\usage{
isDescendant(procPed, a, d)
}
\arguments{
\item{procPed}{pedigree that has been through \code{processPedigree}}

\item{a}{ancestor subject}

\item{d}{descendant subject}
}
\value{
true if d is descended from a
}
\description{
determine if one subject is a descendant of another
}
\keyword{internal}

