% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainNetworkHelper.R
\name{marginalProb}
\alias{marginalProb}
\title{\code{marginalProb} calculates the probability
 that a set of nodes are in given states}
\usage{
marginalProb(net, states)
}
\arguments{
\item{net}{bayesian network from gRain package}

\item{states}{named list of states for each node}
}
\value{
joint probability
}
\description{
\code{marginalProb} calculates the probability
 that a set of nodes are in given states
}
\keyword{internal}
