% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{mendelProbTable}
\alias{mendelProbTable}
\title{Table of mendelian inheritance probabilities}
\usage{
mendelProbTable
}
\description{
Table of mendelian inheritance probabilities
}
\keyword{internal}

