% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculations.R
\name{multipleFamilyPValue}
\alias{multipleFamilyPValue}
\title{probability of sharing of rare variants in a subset of families}
\usage{
multipleFamilyPValue(probs, shared)
}
\arguments{
\item{probs}{sharing probabilities for all families}

\item{shared}{boolean vector describing if all affected subjects
in the family share the variant (TRUE if all share)}
}
\value{
P-value of the exact rare variant sharing test requiring
 sharing by all affected subjects
}
\description{
Computing probability of sharing of rare variants in
 a subset of families where rare variants are seen based on precomputed
 family-specific rare variant sharing probabilities.
}
\details{
All the subsets of families of size equal or inferior to the
 length of not are created, and the joint probability of each such
 subset not sharing a rare variant and the remaining families sharing
 a rare variant is obtained as the product of the family-specific rare
 variant sharing probabilities or its complement. The function then sums
 the pattern probabilities inferior or equal to the probability
 of the observed pattern of the not families not sharing a rare variant
 and the remaining families sharing a rare variant.
}
\examples{
 data(samplePedigrees)
 notSharedFams <- c(15159, 15053, 15157)
 famids <- sapply(samplePedigrees, function(p) p$famid[1])
 shared <- famids \%in\% notSharedFams
 probs <- sapply(samplePedigrees, RVsharing)
 multipleFamilyPValue(probs, shared)
}

