% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculations.R
\name{multipleFamilyPValue_R}
\alias{multipleFamilyPValue_R}
\title{R backend for multipleFamilyPValue calculation}
\usage{
multipleFamilyPValue_R(sharingProbs, observedSharing, minPValue = 0)
}
\arguments{
\item{sharingProbs}{named vector of sharing probabilties, where names
correspond to famid value of pedigree}

\item{observedSharing}{boolean vector describing if all affected subjects
in the family share the variant (TRUE if all share)}

\item{minPValue}{the minimum p-value threshold, once the true p-value is
determined to be less than this, the computation stops and minPValue is
returned - this prevents extremely long computations for extremely small
p-values}
}
\value{
p-value
}
\description{
R backend for multipleFamilyPValue calculation
}
