% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculations.R
\name{multipleVariantPValue}
\alias{multipleVariantPValue}
\title{generalization of multipleFamilyPValue to multiple variants}
\usage{
multipleVariantPValue(snpMat, famInfo, sharingProbs, filter_type = NULL,
  alpha = 0)
}
\arguments{
\item{snpMat}{SnpMatrix}

\item{famInfo}{vector of family id's - same length as rows in share_matrix}

\item{sharingProbs}{vector of subject id's - same length as rows in share_matrix}

\item{filter}{criteria for filtering pvalues}

\item{alpga}{parameter for filter}
}
\value{
vector : p-value for each variant sharing pattern
}
\description{
Computes a p-value for each variant sharing pattern
 across families
}
\details{
For each variant, the families which have all sequenced subjects
 sharing the variant and the families which have some sequenced subjects
 sharing the variant are recorded. These values are passed
 to multipleFamilyPValue
}

