% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculations.R
\name{multipleVariantPValue_R}
\alias{multipleVariantPValue_R}
\title{R backend for multipleVariantPValue calculation}
\usage{
multipleVariantPValue_R(snpMat, famIds, sharingProbs, minorAllele,
  filter = NULL, alpha = 0)
}
\arguments{
\item{snpMat}{SnpMatrix}

\item{sharingProbs}{vector of sharing probabilites, must be a named vector
with famid's for each probability}

\item{minorAllele}{vector specifying the minor allele of each variant}

\item{filter}{criteria for filtering pvalues}

\item{alpha}{parameter for filter}
}
\value{
list of p-values and potential p-values
}
\description{
R backend for multipleVariantPValue calculation
}
