% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingProbabilityCalculations.R
\name{numerProbPattern}
\alias{numerProbPattern}
\title{numerator of sharing probability for a specific sharing pattern of 1 or 2 copies of the RV}
\usage{
numerProbPattern(patternIndex, net, procPed)
}
\arguments{
\item{procPed}{pedigree object that has been process with processPedigree}

\item{gRain}{bayesian network}
}
\value{
numerator value
}
\description{
calculates the numerator of the sharing probability
for a specific sharing pattern of 1 or 2 copies of the RV
}
\keyword{internal}
