% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing.R
\name{oldArgs}
\alias{oldArgs}
\title{check for arguments in v1.7 format}
\usage{
oldArgs(ped, data, dad.id, mom.id)
}
\arguments{
\item{ped}{pedigree object}

\item{data}{old argument}

\item{dad.id}{old argument}

\item{mom.id}{old argument}
}
\value{
if old arguments are provided, a pedigree object is returned,
 otherwise ped is returned
}
\description{
check arguments provided in ... to see if
 the user called RVsharing using a function signature from v1.7, this
 will convert the arguments into a pedigree suitable for the
 signature in version > 2.0
}
\keyword{internal}

