% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree-methods.R
\name{ped2trio}
\alias{ped2trio}
\title{depreciated function}
\usage{
ped2trio(...)
}
\arguments{
\item{...}{arguments to the old function}
}
\value{
none
}
\description{
This function is depreciated with version >= 2.0
 and should not be used.
}
\examples{
tryCatch(ped2trio(), error = function(e) message(e))
}

