% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.esllasso.R
\name{predict.esllasso}
\alias{predict.esllasso}
\title{Make predictions from a "esllasso" object.}
\usage{
\method{predict}{esllasso}(object, newx, ...)
}
\arguments{
\item{object}{Output from the \code{esllasso} function.}

\item{newx}{New data used for prediction}

\item{...}{Additional arguments affecting the predictions produced.}
}
\value{
The predicted responses.
}
\description{
Returns predictions from a fitted
"\code{esllasso}" object.
}
