% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.robustlm.R
\name{predict.robustlm}
\alias{predict.robustlm}
\title{Make predictions from a "robustlm" object.}
\usage{
\method{predict}{robustlm}(object, newx, ...)
}
\arguments{
\item{object}{Output from the \code{robustlm} function.}

\item{newx}{New data used for prediction}

\item{...}{Additional arguments affecting the predictions produced.}
}
\value{
The predicted responses.
}
\description{
Returns predictions from a fitted
"\code{robustlm}" object.
}
