% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.rvs.R
\name{predict.rvs}
\alias{predict.rvs}
\title{Make predictions from a "rvs" object.}
\usage{
\method{predict}{rvs}(object, newx, ...)
}
\arguments{
\item{object}{Output from the \code{rvs} function.}

\item{newx}{New data used for prediction}

\item{...}{Additional arguments affecting the predictions produced.}
}
\value{
The predicted responses.
}
\description{
Returns predictions from a fitted
"\code{rvs}" object.
}
