% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.robustlm.R
\name{print.robustlm}
\alias{print.robustlm}
\title{Print method for a "robustlm" object}
\usage{
\method{print}{robustlm}(x, ...)
}
\arguments{
\item{x}{A "\code{robustlm}" object.}

\item{...}{Additional print arguments.}
}
\value{
print a \code{robustlm} object.
}
\description{
Print the primary elements of the "\code{robustlm}" object.
}
