% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rvs.R
\name{print.rvs}
\alias{print.rvs}
\title{Print method for a "rvs" object}
\usage{
\method{print}{rvs}(x, ...)
}
\arguments{
\item{x}{A "\code{rvs}" object.}

\item{...}{Additional print arguments.}
}
\value{
print a \code{rvs} object.
}
\description{
Print the primary elements of the "\code{rvs}" object.
}
