% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeHelper.R
\name{processPedigree}
\alias{processPedigree}
\title{pre-process a pedigree for calculations}
\usage{
processPedigree(ped, carriers = NULL)
}
\arguments{
\item{ped}{pedigree objevt (S3)}

\item{carriers}{subjects of interest who have the variant}
}
\value{
list containing relevant pedigree info
}
\description{
pre-process a pedigree for calculations
}

