% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloSharingProb.R
\name{runMonteCarlo}
\alias{runMonteCarlo}
\title{\code{runMonteCarlo} run the monte carlo simulation}
\usage{
runMonteCarlo(procPed, founderDist, nSimulations)
}
\arguments{
\item{procPed}{pedigree that has been through processPedigree()}

\item{nSimulations}{number of simulations used in calculation}

\item{founderFunc}{function describing how to sample alleles
for the founders}
}
\value{
sharing probability
}
\description{
\code{runMonteCarlo} run the monte carlo simulation
}
\keyword{internal}
