% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing.R
\docType{methods}
\name{RVsharing}
\alias{RVsharing}
\alias{RVsharing,pedigree-method}
\alias{RVsharing,list-method}
\title{probability of sharing a rare variant among relatives}
\usage{
RVsharing(
  ped,
  carriers = NULL,
  alleleFreq = NA,
  kinshipCoeff = NA,
  nSim = NA,
  founderDist = NULL,
  useAffected = FALSE,
  kinshipOrder = 5,
  splitPed = FALSE,
  useFounderCouples = TRUE,
  distinguishHomo = FALSE,
  ...
)

\S4method{RVsharing}{pedigree}(
  ped,
  carriers,
  alleleFreq,
  kinshipCoeff,
  nSim,
  founderDist,
  useAffected,
  kinshipOrder,
  splitPed,
  useFounderCouples,
  distinguishHomo,
  useUnaffected = FALSE,
  ...
)

\S4method{RVsharing}{list}(
  ped,
  carriers = NULL,
  alleleFreq = NA,
  kinshipCoeff = NA,
  nSim = NA,
  founderDist = NULL,
  useAffected = FALSE,
  kinshipOrder = 5,
  splitPed = FALSE,
  useFounderCouples = TRUE,
  distinguishHomo = FALSE,
  ...
)
}
\arguments{
\item{ped}{S3 pedigree object or a list of pedigree objects}

\item{carriers}{subjects in pedigree that have the variant, if
ped is a list, then this will also be a list of vectors specifying
the carriers in each pedigree}

\item{alleleFreq}{allele frequency among the founders}

\item{kinshipCoeff}{mean kinship coefficient among the founders}

\item{nSim}{number of simulations used in monte carlo calculation}

\item{founderDist}{custom distribution among founders. Only used
when simulating probability with nSim}

\item{useAffected}{a logical value indicating whether to condition on seeing the variant
among the affected subjects in ped instead of the final descendants}

\item{kinshipOrder}{order of the polynomial approximation to the distribtion
of the number of distinct alleles in the founders (d in Bureau et al.).
Must be <= 5}

\item{splitPed}{a logical value indicating whether to split the pedigree in subpedigrees below each founder to enable computations in pedigrees too large to be stored in a single Bayesian network. Cannot be used in conjunction with alleleFreq or kinshipCoef. If useAffected=TRUE, this option produces correct probabilities only if all the affected subjects are non-founders of the pedigree. Erroneous probabilities may be returned if any founder is affected.}

\item{useFounderCouples}{a logical value indicating whether to exploit the interchangeability of the mother and father from founder couples to save computations. Warning! This works only when all founders have only one spouse. Set to FALSE if at least one founder has two or more spouses. Only used when splitPed = TRUE}

\item{distinguishHomo}{a logical value indicating whether to compute distinct probabilities for homozygous and heterozygous variant carrier status}

\item{...}{allows for additional arguments}

\item{useUnaffected}{a logical value indicating whether to include unaffected relatives (pedigree members with affected=0) in the computation of probabilities of rare variant sharing by carriers. The condition that the variant be seen in at least one affected relative remains. Default to FALSE.}
}
\value{
sharing probability between all carriers in pedigree, except when splitPed = TRUE, where a vector of sharing probabilities for all subsets of the carriers is returned, and when distinguishHomo = TRUE, where a vector of sharing probabilities of the (2 to the power ncarriers) configurations of one or two copies of the variant allele among the carriers is returned. In this case, the probabilities are named by binary strings, such that a "0" means 1 copy of the variant allele and a "1" means 2 copies of the variant allele.
}
\description{
computing probability that a rare variant is shared by a
set of subjects in a pedigree using the gRain package
}
\details{
the function RVsharing computes the probability that all subjects
identified as carriers of a rare variant in the vector carriers
(or all final descendants in the pedigree if carriers == NULL) share that
rare variant AND the final descendants not included in carriers do not
carry it, given that the rare variant has been detected in any subject
in the union of the carriers and the final descendants of the pedigree,
when distinguishHomo = FALSE by default.
A final descendant is defined as a subject without descendant in the
pedigree, it it not necessarily in the youngest generation. If carriers
enumerates a subset of pedigree members, the function will then compute
the probability these carriers share the rare variant AND the final
descendants not included in carriers do not carry it based on the above
terms. To control the subset of pedigree members for which sharing of the
rare variant is evaluated, define the members of the subset as affected in
ped and set the option useAffected=TRUE. The only restriction is that all
carriers must be defined as affected. The probability of the (2 to the power
ncarriers) configurations of one or two copies of the variant allele among 
the carriers is instead computed when distinguishHomo = TRUE.
}
\examples{
data("samplePedigrees")
RVsharing(samplePedigrees$firstCousinPair)
}
\references{
Bureau, A., Younkin, S., Parker, M.M., Bailey-Wilson, J.E.,
Marazita, M.L., Murray, J.C., Mangold, E., Albacha-Hejazi, H., Beaty, T.H.
and Ruczinski, I. (2014) Inferring rare disease risk variants based on
exact probabilities of sharing by multiple affected relatives.
Bioinformatics, 30(15): 2189-96, doi:10.1093/bioinformatics/btu198.

Sherman, T., Fu, J., Scharpf, R., Bureau, A., and Ruczinski, I. (2018)
Detection of rare disease variants in extended pedigrees using RVS.
Bioinformatics, 1-3, doi: 10.1093/bioinformatics/bty976
}
