% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing.R
\name{RVsharing}
\alias{RVsharing}
\title{Compute Rare Variant Sharing Probabilities}
\usage{
RVsharing(ped, carriers, alleleFreq, kinshipCoeff, nSim, founderDist, ...)
}
\arguments{
\item{ped}{S3 Pedigree object}

\item{alleleFreq}{allele frequency among the founders}

\item{kinshipCoeff}{mean kinship coefficient among the founders}

\item{nSim}{number of simulations used in monte carlo calculation}
}
\value{
sharing probability between all carriers in pedigree
}
\description{
Calculates rare variant sharing probability between
 carriers in a pedigree given that the variant in seen in one of the
}
\details{

}
\examples{
 data("samplePedigrees")
 RVsharing(samplePedigrees[[1]])
}

