% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing2.R, R/documentation.R
\docType{package}
\name{RVsharing2}
\alias{RVsharing2}
\alias{RVsharing2-package}
\title{\code{RVsharing2} Calculates rare variant sharing probability between
 affected subjects in a pedigree}
\usage{
RVsharing2(ped, alleleFreq, kinshipCoeff, nSimulations)
}
\arguments{
\item{ped}{pedigree object (S3)}

\item{alleleFreq}{frequency of variant present amount the founders}

\item{kinshipCoeff}{relationship between founders}

\item{nSimulations}{number of simulations to run for the monte carlo
approach to calculating sharing probabilities}
}
\value{
sharing probability between all affected subjects in the pedigree
}
\description{
\code{RVsharing2} Calculates rare variant sharing probability between
 affected subjects in a pedigree

RVsharing2
}
\examples{
 data("samplePedigrees")
 RVsharing2(samplePedigrees[[1]])
}

