% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloSharingProb.R
\name{simulatePedigree}
\alias{simulatePedigree}
\title{simulates pedigree given founder states}
\usage{
simulatePedigree(procPed, states)
}
\arguments{
\item{procPed}{pedigree that has been through processPedigree()}

\item{states}{state of each founder (0,1,2 copies of variant)}
}
\value{
states for all subjects in pedigree
}
\description{
#TODO
}
\keyword{internal}

