% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{snpMat}
\alias{snpMat}
\title{SnpMatrix with genotype information from famVCF for fam15157}
\usage{
snpMat
}
\description{
SnpMatrix with genotype information from famVCF for fam15157
}

