% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVgene.R
\name{SnpMatrixToCount}
\alias{SnpMatrixToCount}
\title{convert a list of SnpMatrices to a single matrix in a similiar format
 as LINKAGE except with minor allele counts}
\usage{
SnpMatrixToCount(matList, pedList)
}
\arguments{
\item{matList}{list of SnpMatrices}

\item{pedList}{list of pedigrees}
}
\value{
matrix in LINKAGE format
}
\description{
creates a matrix in LINKAGE format using pedigree information
 from a list of pedigree objects and genotype information from a list of
 SnpMatrices
}
\examples{
 data(samplePedigrees)
 data(snpMat)
 ped <- samplePedigrees$secondCousinTriple
 ex.ped.mat <- SnpMatrixToCount(list(snpMat), list(ped))
}

