% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing.R
\name{checkArgs}
\alias{checkArgs}
\title{check arguments provided to RVsharing for validty}
\usage{
checkArgs(alleleFreq, kinshipCoeff, nSim, founderDist)
}
\arguments{
\item{alleleFreq}{see RVsharing}

\item{kinshipCoeff}{see RVsharing}

\item{nSim}{see RVsharing}

\item{founderDist}{see RVsharing}

\item{...}{see RVsharing}
}
\value{
throws error if arguments invalid
}
\description{
verifies that arguments are valid, throws an error
 if they are not
}
\keyword{internal}

