% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculations.R
\name{convertMatrix}
\alias{convertMatrix}
\title{convert snpMatrix to a list of vectors of sharing}
\usage{
convertMatrix(snpMat, famInfo, minorAllele = NA)
}
\arguments{
\item{snpMat}{SnpMatrix}

\item{famInfo}{data frame containing pedigree, member, father, mother,
sex, affected fields for each sequenced subject}

\item{minorAllele}{vector specifying the minor allele of each variant}
}
\value{
list of boolean vectors indicating sharing pattern for each variant
}
\description{
convert snpMatrix to a list of vectors of sharing
}
