% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainNetworkHelper.R
\name{createNetwork}
\alias{createNetwork}
\title{create bayesian network from processed pedigree}
\usage{
createNetwork(procPed, prior = c(1, 2, 1))
}
\arguments{
\item{procPed}{processed Pedigree object}

\item{prior}{prior on number of alleles for founders}
}
\value{
bayesian network from gRain package
}
\description{
Creates a bayesian network using the gRain package.
 The network is built based on the information in a pedigree object
 that has been processed using \code{processPedigree}.
}
\keyword{internal}

