% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grainNetworkHelper.R
\name{createNetwork}
\alias{createNetwork}
\title{\code{createNetwork} create bayesian network from procPedigree}
\usage{
createNetwork(procPed, prior = c(1, 2, 1))
}
\arguments{
\item{procPed}{processed Pedigree object}

\item{prior}{prior on number of alleles for founders}
}
\value{
bayesian network friom gRain package
}
\description{
\code{createNetwork} create bayesian network from procPedigree
}
\keyword{internal}

