% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculations.R
\name{enrichmentPValue}
\alias{enrichmentPValue}
\title{enrichment p-value acroos multiple families and variants}
\usage{
enrichmentPValue(snpMat, famInfo, sharingProbs, threshold = 0)
}
\arguments{
\item{threshold}{minimum p-value threshold passed to multipleFamilyPValue}
}
\value{
p-value
}
\description{
Computes a p-value for all variants seen across all families
}
\details{
For each variant, the families which have all sequenced subjects
sharing the variant and the families which have some sequenced subjects
sharing the variant are recorded. All unique (family, variant) pairs
are accumulated into a single vector and passed to multipleFamilyPValue
}

