% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{ex.ped.mat}
\alias{ex.ped.mat}
\title{matrix of pedigree information stored in the LINKAGE format}
\usage{
ex.ped.mat
}
\description{
matrix of pedigree information stored in the LINKAGE format
}
