% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedFoundersCorrection.R
\name{inferNumAlleles}
\alias{inferNumAlleles}
\title{most likely number of distinct alleles among founders}
\usage{
inferNumAlleles(phi, nf)
}
\arguments{
\item{phi}{mean estimated kinship coefficient}

\item{nf}{number of founders}
}
\value{
number of distinct alleles
}
\description{
Calculates the most likely number of distinct alleles
among nf founders based on the mean estimated kinship coefficient
}
\keyword{internal}
