% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing2.R
\name{marginalProb}
\alias{marginalProb}
\title{\code{marginalProb} calculates probability that 1) all marginal nodes
 are 0 and 2) all marginal nodes are 1}
\usage{
marginalProb(net, marginalNodes)
}
\arguments{
\item{net}{Bayesian network from gRain package}

\item{marginalNodes}{nodes in the joint-marginal distribution}
}
\value{
two probabilities
}
\description{
\code{marginalProb} calculates probability that 1) all marginal nodes
 are 0 and 2) all marginal nodes are 1
}
\keyword{internal}

