% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloMethods.R
\name{monteCarloSharingProb}
\alias{monteCarloSharingProb}
\title{calculates sharing probability by simulating pedigree outcomes}
\usage{
monteCarloSharingProb(procPed, alleleFreq, kinshipCoeff, nSim, founderDist)
}
\arguments{
\item{procPed}{pedigree that has been through \code{processPedigree}}

\item{alleleFreq}{allele frequency among the founders}

\item{kinshipCoeff}{mean kinship coefficient among the founders}

\item{nSim}{number of simulations used in monte carlo calculation}

\item{founderDist}{custom distribution among founders - only used
when simulating probability with nSim}
}
\value{
sharing probability between all carriers in pedigree
}
\description{
Calculates the same exact probability as RVsharing,
 except uses monte carlo simulation instead of exact computation.
 This method allows for more flexibility in the scenarios considered.
}
\keyword{internal}

