% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloSharingProb.R
\name{monteCarloSharingProb}
\alias{monteCarloSharingProb}
\title{\code{monteCarloSharingProb}}
\usage{
monteCarloSharingProb(procPed, alleleFreq, kinshipCoeff, nSimulations)
}
\description{
\code{monteCarloSharingProb}
}

