% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloSharingProb.R
\name{monteCarloSharingProb}
\alias{monteCarloSharingProb}
\title{\code{monteCarloSharingProb} calculates sharing probability by
 simulating many outcomes of the pedigree}
\usage{
monteCarloSharingProb(procPed, alleleFreq, kinshipCoeff, nSimulations)
}
\arguments{
\item{procPed}{pedigree that has been through processPedigree()}

\item{alleleFreq}{allele frequency among the founders}

\item{kinshipCoeff}{mean kinship coefficient among the founders}

\item{nSimulations}{number of simulations used in calculation}
}
\value{
sharing probability
}
\description{
\code{monteCarloSharingProb} calculates sharing probability by
 simulating many outcomes of the pedigree
}
\keyword{internal}
