% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilies.R
\name{multipleFamilyPValue}
\alias{multipleFamilyPValue}
\title{Probability of sharing of rare variants in a subset of families}
\usage{
multipleFamilyPValue(shared, notShared)
}
\arguments{
\item{shared}{vector of sharing probabilities for families where
all affected subjects share the variant}

\item{notShared}{vector of sharing probabilities for families where
not all affected subjects share the variant}
}
\value{
p-value
}
\description{
Computing probability of sharing of rare variants in
 a subset of families where rare variants are seen based on precomputed
 family-specific rare variant sharing probabilities.
}
\details{
All the subsets of families of size equal or inferior to the
 length of not are created, and the joint probability of each such
 subset not sharing a rare variant and the remaining families sharing
 a rare variant is obtained as the product of the family-specific rare
 variant sharing probabilities or its complement. The function then sums
 the pattern probabilities inferior or equal to the probability
 of the observed pattern of the not families not sharing a rare variant
 and the remaining families sharing a rare variant.
}

