% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculations.R
\name{multipleFamilyPValue}
\alias{multipleFamilyPValue}
\title{Computes a p-value for a single variant across multiple families}
\usage{
multipleFamilyPValue(sharingProbs = NULL, observedSharing, pedList = NULL)
}
\arguments{
\item{sharingProbs}{vector of sharing probabilties}

\item{observedSharing}{vector of T/F indicating which pedigrees have the
variant seen in all carriers}

\item{pedList}{list of pedigree objects}
}
\value{
vector : p-value for each variant sharing pattern
}
\description{
Computes a p-value for a single variant sharing pattern
 across families
}

