% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleFamilyCalculations.R
\name{multipleVariantPValue}
\alias{multipleVariantPValue}
\title{generalization of multipleFamilyPValue to multiple variants}
\usage{
multipleVariantPValue(share_matrix, famid, subjects, peds)
}
\arguments{
\item{share_matrix}{matrix entry [i,j] is 1 if subject i has
variant j, otherwise it is zero}

\item{famid}{vector of family id's - same length as rows in share_matrix}

\item{subjects}{vector of subject id's - same length as rows in share_matrix}

\item{peds}{list of pedigrees referenced in famid}
}
\value{
vector : p-value for each variant sharing pattern
}
\description{
Computes a p-value for each variant sharing pattern
 across families
}
\details{
For each variant, the families which have all sequenced subjects
 sharing the variant and the families which have some sequenced subjects
 sharing the variant are recorded. These values are passed
 to multipleFamilyPValue
}

