% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing2.R
\name{oneFounderSharingProb}
\alias{oneFounderSharingProb}
\title{\code{oneFounderSharingProb} calculate sharing prob assuming one founder
 introduces the variant}
\usage{
oneFounderSharingProb(ped)
}
\arguments{
\item{ped}{pedigree S3 Object}
}
\value{
sharing probability
}
\description{
\code{oneFounderSharingProb} calculate sharing prob assuming one founder
 introduces the variant
}
\keyword{internal}

