% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingProbabilityCalculations.R
\name{oneFounderSharingProb}
\alias{oneFounderSharingProb}
\title{calculate sharing probability in basic case}
\usage{
oneFounderSharingProb(procPed)
}
\arguments{
\item{procPed}{pedigree that has been through processPedigree()}
}
\value{
sharing probability
}
\description{
Assume that only one founder can introduce the variant to 
 the pedigree. Condition on each founder and sum over all resulting
 probabilities.
}
\keyword{internal}

