% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingProbabilityCalculationsSplitting.R
\name{oneFounderSharingProbSplitting}
\alias{oneFounderSharingProbSplitting}
\title{calculate sharing probability in basic case}
\usage{
oneFounderSharingProbSplitting(procPed)
}
\arguments{
\item{procPed}{pedigree that has been through processPedigree()}
}
\value{
sharing probability
}
\description{
Assume that only one founder can introduce the variant to 
the pedigree. Condition on each founder and sum over all resulting
probabilities.
}
\keyword{internal}
