% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree-methods.R
\name{ped2trio}
\alias{ped2trio}
\title{depreciated function}
\usage{
ped2trio(...)
}
\arguments{
\item{...}{arguments to the old function}
}
\description{
This function is depreciated with version >= 2.0
 and should not be used.
}

