% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree-methods.R
\docType{methods}
\name{processPedigree}
\alias{processPedigree}
\alias{processPedigree,pedigree-method}
\title{extract useful information from a pedigree}
\usage{
processPedigree(ped, carriers = NULL)

\S4method{processPedigree}{pedigree}(ped, carriers = NULL)
}
\arguments{
\item{ped}{pedigree object (S3)}

\item{carriers}{subjects in which the rare variant is seen}
}
\value{
list containing relevant pedigree info
}
\description{
Extract key information from a pedigree object, which makes
 subsequent computations much easier.
}
\examples{
 data(samplePedigrees)
 processPedigree(samplePedigrees$firstCousinPair)
}

