% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigreeHelper.R
\name{processPedigree}
\alias{processPedigree}
\title{\code{processPedigree} pre-process a pedigree for calculations}
\usage{
processPedigree(ped)
}
\arguments{
\item{ped}{pedigree objevt (S3)}
}
\value{
list containing relevant pedigree info
}
\description{
\code{processPedigree} pre-process a pedigree for calculations
}
\keyword{internal}

