% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedFoundersCorrection.R
\name{relatedFoundersCorrection}
\alias{relatedFoundersCorrection}
\title{make the neccesary correction for when founders have a non-zero
 kinship coefficient}
\usage{
relatedFoundersCorrection(nf, kinshipCoeff)
}
\arguments{
\item{nf}{number of founders}

\item{kinshipCoeff}{mean kinship coefficient among all founders}
}
\value{
weight used in probability calculation
}
\description{
make the neccesary correction for when founders have a non-zero
 kinship coefficient
}
\keyword{internal}

