% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloSharingProb.R
\name{runMonteCarlo}
\alias{runMonteCarlo}
\title{run the monte carlo simulation}
\usage{
runMonteCarlo(procPed, founderDist, nSim)
}
\arguments{
\item{procPed}{pedigree that has been through \code{processPedigree}}

\item{founderDist}{custom distribution among founders - only used
when simulating probability with nSim}

\item{nSim}{number of simulations used in monte carlo calculation}
}
\value{
sharing probability between all carriers in pedigree
}
\description{
Given a number of simulations and a distribution
 of variants in the founders, this function simulates possbile
 outcomes of the pedigree and returns a sharing probability.
}
\keyword{internal}

