% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monteCarloMethods.R
\name{runMonteCarlo}
\alias{runMonteCarlo}
\title{run the monte carlo simulation}
\usage{
runMonteCarlo(procPed, founderDist, nSim, runParallel = FALSE)
}
\arguments{
\item{procPed}{pedigree that has been through \code{processPedigree}}

\item{founderDist}{custom distribution among founders - only used
when simulating probability with nSim}

\item{nSim}{number of simulations used in monte carlo calculation}

\item{runParallel}{run simulation in parallel}
}
\value{
sharing probability between all carriers in pedigree
}
\description{
Given a number of simulations and a distribution
 of variants in the founders, this function simulates possbile
 outcomes of the pedigree and returns a sharing probability.
}
\details{
If the number of simulations is greater than 20,000
 then the computation is done in parallel (as long as the package
 parallel is available)
}
\keyword{internal}

