% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing.R
\docType{methods}
\name{RVsharing}
\alias{RVsharing}
\alias{RVsharing,list-method}
\alias{RVsharing,pedigree-method}
\title{probability of sharing a rare variant among relatives}
\usage{
RVsharing(ped, carriers, alleleFreq, kinshipCoeff, nSim, founderDist,
  useAffected = FALSE, ...)

\S4method{RVsharing}{pedigree}(ped, carriers, alleleFreq, kinshipCoeff, nSim,
  founderDist, useAffected = FALSE, ...)

\S4method{RVsharing}{list}(ped, carriers, alleleFreq, kinshipCoeff, nSim,
  founderDist, useAffected = FALSE, ...)
}
\arguments{
\item{ped}{S3 pedigree object or a list of pedigree objects}

\item{carriers}{subjects in pedigree that have the variant, if
ped is a list, then this will also be a list of vectors specifying
the carriers in each pedigree}

\item{alleleFreq}{allele frequency among the founders}

\item{kinshipCoeff}{mean kinship coefficient among the founders}

\item{nSim}{number of simulations used in monte carlo calculation}

\item{founderDist}{custom distribution among founders - only used
when simulating probability with nSim}

\item{useAffected}{allows the user to condition on seeing the variant
among the affected subjects instead of the final descendants}

\item{...}{allows for arguments in the style of v1.7}
}
\value{
sharing probability between all carriers in pedigree
}
\description{
computing probability that a rare variant is shared by a
 set of subjects in a pedigree using the gRain package
}
\details{
the function RVsharing computes the probability that all subjects
 identified as carriers of a rare variant in the vector carriers
 (or all final descendants in the pedigree if carriers == NULL) share that
 rare variant AND the final descendants not included in carriers do not
 carry it, given that the rare variant has been detected in any subject
 in the union of the carriers and the final descendants of the pedigree.
 A final descendant is defined as a subject without descendant in the
 pedigree, it it not necessarily in the youngest generation. If carriers
 enumerates a subset of pedigree members, the function will then compute
 the probability these carriers share the rare variant AND the final
 descendants not included in carriers do not carry it based on the above
 terms. To obtain the probability that a set of pedigree members carry a
 rare variant given it was seen in any of the set members (ignoring the
 carrier status of final descendants not in the set), the pedigree must be
 trimmed of the other final descendants before calling RVsharing.
}
\examples{
 data("samplePedigrees")
 RVsharing(samplePedigrees$firstCousinPair)
}

