% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing2.R, R/documentation.R
\docType{package}
\name{RVsharing2}
\alias{RVsharing2}
\alias{RVsharing2}
\alias{RVsharing2-package}
\title{\code{RVsharing2} Calculates rare variant sharing probability between
 carriers in a pedigree given that the variant in seen in one of the}
\usage{
RVsharing2(ped, carriers, alleleFreq, kinshipCoeff, nSimulations, founderDist)
}
\arguments{
\item{ped}{S3 Pedigree object}

\item{alleleFreq}{allele frequency among the founders}

\item{kinshipCoeff}{mean kinship coefficient among the founders}

\item{nSimulations}{number of simulations used in monte carlo calculation}
}
\value{
sharing probability between all carriers in pedigree
}
\description{
\code{RVsharing2} Calculates rare variant sharing probability between
 carriers in a pedigree given that the variant in seen in one of the

RVsharing2
}
\examples{
 data("samplePedigrees")
 RVsharing2(samplePedigrees[[1]])
}
