% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingProbabilityCalculations.R
\name{twoFounderSharingProb}
\alias{twoFounderSharingProb}
\title{sharing probability when founder pair introduces variant}
\usage{
twoFounderSharingProb(
  procPed,
  kinshipCoeff,
  kinshipOrder,
  distinguishHomo = FALSE
)
}
\arguments{
\item{procPed}{pedigree that has been through processPedigree()}

\item{kinshipCoeff}{mean kinship coefficient among the founders}

\item{kinshipOrder}{order of the polynomial approximation to the distribtion}

\item{distinguishHomo}{a logical value indicating whether to compute distinct probabilities for homozygous and heterozygous variant carrier status
of the number of distinct alleles in the founders (d in Bureau et al.).
Must be <= 5}
}
\value{
sharing probability
}
\description{
In the case of relatedness among founders, assume that up
to two founders could introduce the variant and condition on all possible
pairs.
}
\keyword{internal}
