% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharingProbabilityCalculation.R
\name{twoFounderSharingProb}
\alias{twoFounderSharingProb}
\title{calculate the sharing probability when two founders can introduce variant}
\usage{
twoFounderSharingProb(procPed, kinshipCoeff)
}
\arguments{
\item{procPed}{pedigree that has been through processPedigree()}

\item{kinshipCoeff}{mean kinship coefficient among the founders}
}
\value{
sharing probability
}
\description{
In the case of relatedness among founders, assume that up
 to two founders could introduce the variant and condition on all possible
 pairs.
}
\keyword{internal}

