% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVsharing2.R
\name{twoFounderSharingProb}
\alias{twoFounderSharingProb}
\title{\code{twoFounderSharingProb} calculate the sharing probability assuming
 that at most two founders introduce the variant, accounting for 
 relatedness among founders}
\usage{
twoFounderSharingProb(procPed, kinshipCoeff)
}
\arguments{
\item{procPed}{pedigree that has been through processPedigree()}

\item{kinshipCoeff}{mean kinship coefficient among the founders}
}
\value{
sharing probability
}
\description{
\code{twoFounderSharingProb} calculate the sharing probability assuming
 that at most two founders introduce the variant, accounting for 
 relatedness among founders
}
\keyword{internal}

